+++ console/src/components/App/App.js
@@ -12,7 +12,6 @@ import globals from '../../Globals';
 import styles from './App.scss';
 
 import { theme } from '../UIKit/theme';
-import { exportMetadata } from '../../metadata/actions';
 
 export const GlobalContext = React.createContext(globals);
 
@@ -38,17 +37,10 @@ const App = ({
       telemetry.console_opts &&
       !telemetry.console_opts.telemetryNotificationShown
     ) {
-      dispatch(telemetryNotificationShown());
       dispatch(showTelemetryNotification());
+      dispatch(telemetryNotificationShown());
     }
   }, [dispatch, telemetry]);
-
-  React.useEffect(() => {
-    console.log('fetch metadata');
-    // no sure yet, if this is the best place
-    // make it being triggered on each metadata update
-    dispatch(exportMetadata());
-  }, []);
 
   let connectionFailMsg = null;
   if (connectionFailed) {